+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- (VANAGON-157) Make curl calls fail on error
 
 ## [0.15.29] - released on 2019-09-25
 ### Changed